requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-groq
fastapi
uvicorn[standard]
markdown
email-validator
sendgrid
